misc(events_processor): Add config to consume dead letter queue events (#4602)
## Context

Today, when an event enrichment failed in the events processor service,
if the error is considered "final" meaning that it cannot be reprocessed
automatically, the event is pushed to the dead letter queue topic with
some extra data (error code, message and failed time).

This topic is configured to store messages forever and nothing is
consuming it.

## Description

This PR adds the local config to consume the dead letter queue topic and
store the transformed data in the `events-dead-letter` clickhouse table