feat(netsuite): Introduce `currencies` mapping (#4519)
## Context

With Netsuite integration, we need to map currencies to their netsuite
objects.

## Description

This PR introduces a new mapping type `currencies`. It's only used for
Netsuite.

We save the mapping as a hashmap because it's best way to store mapping
😄 The mapping is meant to be very rarely edited but used a lot.

Usage:
```rb
payload[:currency] = mapping_model.currencies[invoice.currency]
```

Next PR will start using this mapping to update integration payload.

### Validation

The model validates the format of currencies hashmap. Create and Update
services already handles `ActiveRecord::InvalidRecord`

GraphQL isn't really meant to use free form hashmaps so we use an array
of hashmap. This PR introduces a custom validator to ensure that a "key"
is only used once. I wanted this in the past for threshold for instance
so I made it configurable.

In this case, we cannot return an error code like we usually do. Notice
that the frontend should already make this duplication impossible, but
just in case, the backend will also throw an error.

```rb
# GQL
[ {currency_code: "EUR", external_currency_code: "2"} ]

# Model
{"EUR" => "2"}
```