chore(profiling): Allow to profile Sidekiq jobs in dev environment (#4597)
## Context

It is currently cumbersome to properly debug performance issues related to a Sidekiq job.

There is a profiling feature using [Vernier](https://github.com/jhawthorn/vernier) in Sidekiq 8 but there are 2 caveats:

1. We are still using Sidekiq 7
2. It is not compatible with `ActiveJob`

## Description

This adds a Sidekiq middleware that is included only in dev environment if the `SIDEKIQ_PROFILING_ENABLED` env is set to `true`. This middleware will profile the job using [Vernier](https://github.com/jhawthorn/vernier) and generate profiling results per job.

The profiling results are in JSON format and can be analyzed using the [`profile-viewer`](https://rubygems.org/gems/profile-viewer) gem.