feat(payment): Add error_code (#4634)
Extracted from https://github.com/getlago/lago-api/pull/4623

This introduce a new columns on payment to save the provider error_code.
This is needed for the 3ds retry but I think it's generally useful, even
if we go a different route with 3ds.

Now that I write these line, I'm thinking `provider_error_code` is also
a good name.