misc(clickhouse): Remove aggregated events (#4671)
## Context

Clickhouse pre-aggregation was added in the application to pre-compute
aggregation data for billing and current usage allowing some live-time
usage features.

Because of some new priority this feature is now on old.

## Description

This PR clean the code related to the `events_aggregated` clickhouse
table as it is not in use today and we don't want to keep maintaining
dead code.

The `events_enriched_expanded` logic is kept as the enriched records
will be used to improve the billing and current usage computation.