feat(issuing-date-5) Finalize invoices based on expected_finalization_date (#4667)
## Roadmap Task

👉
[Spec](https://www.notion.so/getlago/Spec-Define-the-issuing-date-preferences-of-subscription-invoices-2a0ef63110d2807d860cc22ced334bfb)
👉 [Dive
In](https://www.notion.so/getlago/BE-Dive-In-Define-the-issuing-date-preferences-of-subscription-invoices-2a3ef63110d2800e8620fb5687f0edf9?d=2a3ef63110d2808f96fd001cbc71298b#2a3ef63110d2807abf14f498b984ac57)

## Context

This PR is part of the Invoice Issuing Date preferences updates.

With the new issuing date adjustment setting, `issuing_date` might skip
using `invoice_grace_period` (when the adjustment set to "keep_anchor").
But we still need to factor in `invoice_grace_period` when finalizing
invoices.

## Description

This PR adds a new field, `expected_finalization_date`, which stores the
date when a draft invoice should be auto-finalized. It's basically:
```
expected_finalization_date = invoice creation date + invoice_grace_period.days
```

The `expected_finalization_date` field is mostly for internal use. We
only expose it in GraphQL so we can show the right message about when a
draft invoice will be finalized. It's not available in the REST API or
in the SDK clients.