feat(event): Add new filter service for billing period (#4676)
## Context

This PR is part of the "pre-filter" events feature.

## Description

The goal of these changes is to extract the logic used to check which
billable metrics where reached by events on a specific billing period
into a new `Events::BillingPeriodFilterService`. This service will be
called by all services performing events aggregations (namely
`Invoices::CalculateFeesService` and `Invoices::CustomerUsageService`).

The new service returns a list of `charge_id` rather than a list of
billable metric codes. It will be updated later to also returns a list
of `charge_filter_ids` allowing the services to perform the aggregation
for filters only when required.