fix(forecasted-usage): Fallback to GraduatedService when we are forecasting amounts for ProratedGraduatedService (#4692)
## Context

Prorated calculations considers event data that we don't have when
forecasting units on the data side, so we fall back to a default
approach of using `GraduatedService` for that.

## Description

Added a default param to the charge model class that checks if the
factory has an `aggregator`. When we are forecasting, we do not have
this `aggregator`, so it fallback to the `GraduatedService`.