misc(webhook): Sent invoice.payment_failure when retry without payment provider (#4691)
## Context

The `invoice.created` webhook was recently removed from the payment
retry flow, because the invoice was not really created.

However, some customer have built a flow around this webhook message.
They are collecting payments outside of Lago, but they are relying on
the `invoice.created` webhook that was sent when an invoice was resent
for collection.

As in this case, no payment method is attached to the customers, the
payment retry on an an invoice does not trigger any webhook message.

## Description

This PR make sure we send the existing `invoice.payment_failure` webhook
in this case