fix(plans): Fix plan creation issues
## Context

When creating a plan with charges, two bugs are currently present:

- Sending a `charge` without `billable_metric_id` will result in an `undefined method 'custom_agg?' for nil (NoMethodError)` error.
- Sending a `null` or empty `values` for a charge filter will result in an `undefined method 'each' for nil:NilClass (NoMethodError)` error.

## Description

The first bug occured because `Plans::CreateService` had a private `create_charge` method that didn't properly check that the `charge` was valid (billable metric present) before proceeding to create it. The charge creation logic was already present in `Charges::CreateService` with proper validations. This replaces the inline charge creation logic in `Plans::CreateService` with a
call to `Charges::CreateService`.

The second bug occurred because the `ChargeFilters::CreateOrUpdateBatchService` service assumed that `filter_params[:values]` would always be present and iterable.