fix(webhook): Discard failed deserialization (#4699)
## Context

This PR follows https://github.com/getlago/lago-api/pull/4693
When a webhook endpoint is deleted, all related webhooks are deleted as
well because of the "dependent: :delete_all" defined in the relation. If
some webhooks are waiting for delivery while an endpoint is being
removed, a `ActiveJob::DeserializationError` is raised by ActiveJob
because the global id does not match an existing record.

## Description

This PR is adding a retry and discard approach to fix this issue. 3
retry attempts are made to ensure we are not in a "not yet commit
database transaction" scenario, and in case of a new error, it is simply
logged, reducing the noise in the Sidekiq dead queue