feat(usage_threshold): Attach UsageThreshold to subscription model (#4675)
## Context

Overriding UsageThreshold on a subscription, will no longer create a
plan override.

## Description

This PR allows usage_thresholds to be attached directly to a
subscription, not just a plan. This is similar to
`Entitlement::Entitlement`.

This is only the ground work to update the model and database. Nothing
is attached to subscription yet.