fix(integration): Fix customer integration creation concurrency (#4694)
## Context

It may happen that a customer was updated in a short time after it was created. As creating integration customers is a long operation, by the time we receive the update we still haven't created the integration customer. Therefore we will schedule a second `IntegrationCustomers::CreateJob`.

## Description

This fixes it by adding job uniqueness until executed. Note that we kept the `integration_customer_params` in the lock key arguments to ensure we still raise an error if the customer update changed the integration customer settings.

I also used this change to rename `IntegrationCustomers::CreateOrUpdateService` to `IntegrationCustomers::CreateOrUpdateBatchService` for consistency.