feat(multiple-payment-methods): update attach cs to invoice logic (#4673)
## Context

With multiple payment methods feature, it will be possible to attach
payment method to different resources like subscription, wallet, wallet
transaction, or one-off invoice

## Description

This PR prepares logic that attaches custom sections on the invoice.

Before custom sections were copied from customer or billing entity to
the invoice. To support new feature, the logic needs to be extended so
that it support custom sections from other resources as well