misc(clickhouse): Remove wrong limit by (#4708)
## Context

Attempts to handle clickhouse deduplication were made using the
`limit_by(1` scope.
It was working with `ActiveRecord` scopes within
`Events::Stores::ClickhouseStore#events` method, without deduplicating
on both `transaction_id` and `timestamp`, but only on `transaction_id`
which is not the behavior documented on Lago's doc.

Most of the aggregation query does not rely on the `events` method, but
on the `events_sql` one, written using `arel`. In this case, the
`limit_by` is used, but is actually doing nothing. Last attempt to fix
it leads us to a MAJOR bug in production
(https://github.com/getlago/lago-api/pull/4418)

## Description

This PR is removing this wrong `limit_by`. It will be replaced by a
better deduplication approach