Feat(fixed_charge): add missing fixed_charge boundaries for invoice_subscriptions (2) (#4468)
## Context

This is a second part of preparation to fixed_charge_fee creation for an
invoice, it includes filling fixed_charge_boundaries for
invoice_subscriptions

## Description

- fill fixed_charge_fee boundaries when creating invoice_subscriptions
- add tests for invoice boundaries. Notice that there are only ~90 lines
of changes, all the rest are tests. Also, some of the introduced in this
PR methods are changed in the next PRs. these methods do not affect
existing logic and we'll check them at the end of the feature

---------

Co-authored-by: Ancor Cruz <hello@ancorcruz.com>