chore(seeds): Add pricing unit to seeds (#4716)
## Context

When testing, manually or automated, we may want to include scenarios with pricing unit. For instance, current usage has different payloads depending on whether pricing unit is applied or not.

There's currently no API to create pricing units so it cannot be automated and it has to be done with the UI. 

## Description

This adds a pricing unit creation as part of the seeds.