fix(seeds): Fix seed webhook in staging env (#4722)
## Context

The seeds currently create a webhook endpoint set to `http://webhook/` which expected to be the [`webhook-tester`](https://github.com/tarampampam/webhook-tester) service. This service is set in the developer docker compose.

The issue is that this service is not defined in staging, causing webhooks to fail and use Sidekiq resources unnecessarily.

## Description

This updates the seeds to only define the webhook endpoint in development environment.