fix(seeds): Fix wallet creation in seeds (#4717)
## Context

The `Wallets::CreateService` was deriving `organization_id` both from `result.current_customer.organization_id` and `params[:organization_id]`, which caused the seeds to silently fail to create wallet transactions because the `organization_id` wasn't provided.

## Description

This refactors the services to require `organization_id` as an explicit parameter and fixes the seeds.