misc(clickhouse): Add service to remove duplicated events (#4725)
## Context

An issue has been raised with the Clickhouse implementation, leading to
a risk of duplicated events when the same transaction_id/timestamp are
sent multiple time.

## Description

A fix to allow a full deduplication is ongoing but it's not ready yet.
In the meantime, this PR is adding a small utility service to allow a
manual cleanup of all duplicated events for a given subscription on a
given billing period.

```ruby
Events::Stores::Clickhouse::CleanDuplicatedService.call!(subscription:)
```

It will remove all duplicated events except the last received one
(ordered by `enriched_at`)