feat(fixed-charges): Include units as criteria to consider as override (#4739)
Before we only consider fixed charges to be overriden if charge model
and properties are different, however, this change include units also as
a condition to be considered an override.

This is, if two fixed charges has same charge model and properties but
different units, then these have different properties and should be
considered an override when checking for parent and child fixed charge.

Then;
Fixed charge units only updates when no cascade or cascade and eq props,
this is charge model, properties and units matching.