feat(fixed-charges) GraphQL: expose fixed charge boundaries for InvoiceSubscription
Currently frontent uses charge boundaries from invoice subscription to
properly render the fees in the right sections depending on its
boundaries. However, InvoiceSubscription object type does not expose
fixed charge and in advance fixed charge boundaries.

This change adds these boundaries to InvoiceSubscription in order to
ease the UI rendering the fees in the right sections.

Ideally, this information should be rather extracted from the fee object
boundaries (from/to datetime) that should have the right set of
boundaries based on the fee type and in arrears / in advance
configuration.