feat(multiple-payment-methods): add api support for payment method on subscription and wallet transaction (#4709)
## Context

With `multiple payment methods` feature, it will be possible to attach
multiple payment methods on customer, but also attach payment method on
resources like subscription, wallet, recurring rules or wallet
transaction.

## Description

This PR is missing part that just expose needed params for subscription
and wallet transaction. The logic itself is already covered in the
services and GQL part is already updated.