feat(multiple-payment-methods): expose payment method details in API for wallet and recurring rules (#4707)
## Context

With `multiple payment methods` feature, it will be possible to attach
multiple payment methods on customer, but also attach payment method on
resources like subscription, wallet, recurring rules or wallet
transaction.

## Description

This PR is missing part that just expose needed params for wallet and
recurring rule. The logic itself is already covered in the services and
GQL part is already updated