misc(moneyhash): Refact webhook handling (#4762)
## Context

Today, payment providers implementation is a bit messy in the code base.
The code is spread between many services and a lot of logic is
duplicated making it hard to maintain. It also prevents easy integration
of new payments providers.

Many improvements have already been made to try to improve the
situation, but it is far from terminated.

## Description

This PR extracted
`::PaymentProviders::MoneyhashService.new.handle_event` method into
`::PaymentProviders::Moneyhash::HandleEventService.call` to follow the
pattern already in place for all other payment providers

NOTE: nothing was changed in the event handler itself. It could probably
be improved as well, but it is outside of the scope of this pull request